Fixing WHERE clauses
Also determining that the B+ data structure itself cannot be reliably executed in parallel. However, all operations using it can be. For that reason, it was renamed and pulled out of the individual engine files. Documentation was quickly updated with this change, will let the team know in person also. Additionally, fixed some of the WHERE clause statements so it seems to work with the current implementation. Now, I will tie in pretty printing and begin working on the QPESeq.c file to get a viable MVP for sequential stuff before continuing.